Resolve some of the warnings in Windows build (#571)
This is a preparation compile with `/WX` flag.

+++ test/TopologicalSortTest.cpp
@@ -96,7 +96,7 @@ TEST(TopologicalSortTest, test_3) {
   ASSERT_EQ(res.nodesInTopoOrder.size(), 8);
 
   // check topological order of nodes
-  std::unordered_map<CXXGraph::id_t, int> nodeToOrder;
+  std::unordered_map<CXXGraph::id_t, size_t> nodeToOrder;
   for (size_t i = 0; i < res.nodesInTopoOrder.size(); ++i) {
     nodeToOrder[res.nodesInTopoOrder[i].getId()] = i;
   }
